cli/pyproject.toml
[project]
name = "cli"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "httpx>=0.28.1",
    "langchain>=0.3.27",
    "langchain-google-genai>=2.1.12",
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.44",
    "sqlmodel>=0.0.27",
    "typer>=0.20.0",
    # Pin orjson to a version with Windows cp312 wheels to avoid Rust build
    "orjson==3.10.7",
    "beautifulsoup4>=4.14.2",
    "readability-lxml>=0.8.4.1",
    "google-search-results>=2.4.2",
]
